superstore sales
Noteworthy Insights:

The total sales for the year amounted to 0.72 million, with the lowest sales recorded in 2016, totaling 0.46 million.
The "Standard Class" shipment mode outperformed both "First Class" and "Second Class."
The "Consumer" segment led in total category revenue, doubling that of the "Home Office" segment.
The Canon image CLASS IMAGE 2200 advanced copier emerged as the highest-selling product, but interestingly, it was not popular in Florida. Furthermore, the categories of Technology, Furniture, and Home Supplies demonstrated similar sales distribution.